+++ CHANGELOG.md
@@ -8,11 +8,22 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 ## Prerelease
 
+### Adds
+
+- Adds a changelog to the story pages for the `Heading` component.
+
+### Updates
+
+- Updates the `Heading` component to set the `aria-roledescription` value as `"subtitle"` (a more familiar and recognizable value) for the `overline` element.
+- Updates the `FeedbackBox` and `NewsletterSignup` components to set the `tabindex` on the internal elements to `"-1"`.
+
 ### Deprecates
 
 - Deprecates the `"secondary"` and `"fiftyFifty"` variants of the `Hero` component.
 
 ## 2.1.0 (October 18, 2023)
+
+### Adds
 
 - Adds the `ComponentChangelogTable` component.
 - Adds a changelog to the story pages for the `DatePicker`, `FeedbackBox`, `Hero`, `Slider`, and `TextInput` components.